chore(pie-cookie-banner): DSW-1644 move definition to bypass TS error
