+++ packages/components/pie-cookie-banner/src/defs.ts
@@ -4,8 +4,6 @@ import {
     TemplateResult,
 } from 'lit';
 
-import defaultLocale from '../locales/en-gb.json';
-
 export interface CookieBannerLocale {
     banner: {
         title: string;
@@ -147,10 +145,3 @@ export interface CustomTagEnhancers {
 
 export type DefaultProps = ComponentDefaultPropsGeneric<CookieBannerProps, 'hasPrimaryActionsOnly' | 'defaultPreferences' | 'locale' | 'cookieStatementLink' | 'cookieTechnologiesLink'>;
 
-export const defaultProps: Readonly<DefaultProps> = {
-    hasPrimaryActionsOnly: false,
-    defaultPreferences: {},
-    locale: defaultLocale,
-    cookieStatementLink: '',
-    cookieTechnologiesLink: '',
-};
