+++ packages/components/pie-cookie-banner/src/index.ts
@@ -15,6 +15,8 @@ import { PieSwitch } from '@justeattakeaway/pie-switch';
 import { defineCustomElement, dispatchCustomEvent } from '@justeattakeaway/pie-webc-core';
 /* eslint-enable import/no-duplicates */
 
+import defaultLocale from '../locales/en-gb.json';
+
 import styles from './cookie-banner.scss?inline';
 import {
     CookieBannerProps,
@@ -27,13 +29,24 @@ import {
     type PreferenceIds,
     type CookieBannerLocale,
     type CustomTagEnhancers,
-    defaultProps,
+    DefaultProps,
 } from './defs';
 
 import { localiseText, localiseRichText } from './localisation-utils';
 
 // Valid values available to consumers
 export * from './defs';
+
+// TODO: Move this object to defs.ts
+// It can be moved as soon as we manage to figure a safe way to avoid the error TS2821
+// Import assertions are only supported when the '--module' option is set to 'esnext' or 'nodenext'
+export const defaultProps: Readonly<DefaultProps> = {
+    hasPrimaryActionsOnly: false,
+    defaultPreferences: {},
+    locale: defaultLocale,
+    cookieStatementLink: '',
+    cookieTechnologiesLink: '',
+};
 
 const componentSelector = 'pie-cookie-banner';
